refactor(event cache): move the `Deduplicator` instance to `RoomEventCacheState`
The `RoomEvents` doesn't hold the `Deduplicator` instance now, it's the
role of the `RoomEventCacheState`. This slightly simplifies the code, in
a few cases.
